+++ src/routes/c/[name]/+page.svelte
@@ -113,7 +113,7 @@
         />
     </div>
 
-    <div class="mt-[-8px]">
+    <div class="mt-[8px]">
         <CommunityCard community_view={data.community.community_view} moderators={data.community.moderators}/>
     </div>
 </div>